% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmtf_gradient.R
\name{cmtf_gradient}
\alias{cmtf_gradient}
\title{Calculate gradient of CMTF model.}
\usage{
cmtf_gradient(x, Z)
}
\arguments{
\item{x}{Vectorized parameters of the CMTF model.}

\item{Z}{Z object as generated by \code{\link[=setupCMTFdata]{setupCMTFdata()}}.}
}
\value{
Vectorized gradient of the CMTF model.
}
\description{
Calculate gradient of CMTF model.
}
\examples{
A = array(rnorm(108*2), c(108, 2))
B = array(rnorm(100*4), c(100, 4))
C = array(rnorm(10*4), c(10, 4))

df1 = reinflateTensor(A, B[,1:2], C[,1:2])
df2 = reinflateTensor(A, B[,3:4], C[,3:4])
datasets = list(df1, df2)
modes = list(c(1,2,3), c(1,4,5))
Z = setupCMTFdata(datasets, modes, normalize=FALSE)

result = cmtf_opt(Z, 1, max_iter=2) # unoptimized CMTF model
f = cmtf_gradient(result$par, Z)
}
