% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_degeneracy.R
\name{cv_degeneracy}
\alias{cv_degeneracy}
\title{Title}
\usage{
cv_degeneracy(
  Z,
  Y,
  numComponents = 1:5,
  alpha = 1,
  beta = rep(0.001, length(Z$object)),
  epsilon = 1e-08,
  pis = seq(0.05, 1, 0.05),
  mu = 1e-06,
  initialization = "random",
  method = "CG",
  cg_update = "HS",
  line_search = "MT",
  max_iter = 10000,
  max_fn = 10000,
  abs_tol = 1e-10,
  rel_tol = 1e-10,
  grad_tol = 1e-10,
  nstart = 100,
  numCores = parallel::detectCores()
)
}
\arguments{
\item{Z}{Combined dataset and mode object as produced by \code{\link[=setupCMTFdata]{setupCMTFdata()}}.}

\item{Y}{Dependent variable (regression part).}

\item{numComponents}{Vector of integer values containing the number of components F to check.}

\item{alpha}{Scalar penalizing the components to be norm 1 (default 1).}

\item{beta}{Vector of penalty values for each dataset, penalizing the lambda terms (default 1e-3).}

\item{epsilon}{Scalar value to make it possible to compute the partial derivatives of lambda (default 1e-8).}

\item{pis}{Vector of pi values to check.}

\item{mu}{Ridge term parameter for calculation of the regression coefficients rho (default = 1e-6).}

\item{initialization}{Initialization, either "random" (default) or "nvec" for numComponents components of the concatenated data using svd. Ignored and uses NPLS based initialization if pi=0.}

\item{method}{Optimization method to use (default = "CG", the conjugate gradient). See \code{\link[mize:mize]{mize::mize()}} for other options.}

\item{cg_update}{Update method for the conjugate gradient algorithm, see \code{\link[mize:mize]{mize::mize()}} for the options (default="HS", Hestenes-Steifel).}

\item{line_search}{Line search algorithm to use, see \code{\link[mize:mize]{mize::mize()}} for the options (default="MT", More-Thuente).}

\item{max_iter}{Maximum number of iterations.}

\item{max_fn}{Maximum number of function evaluations.}

\item{abs_tol}{Function tolerance criterion for convergence.}

\item{rel_tol}{Relative function tolerance criterion for convergence.}

\item{grad_tol}{Absolute tolerence for the l2-norm of the gradient vector.}

\item{nstart}{Number of models to produce (default 1). If set higher than one, the package will return the best fitted model.}

\item{numCores}{Number of cores to use (default 1). If set higher than one, the package will attempt to run in parallel.}
}
\value{
List object containing "models": all models, "settings": pi and F used to create the models.
}
\description{
Title
}
\examples{
set.seed(123)

A = array(rnorm(108*2), c(108, 2))
B = array(rnorm(100*2), c(100, 2))
C = array(rnorm(10*2), c(10, 2))
D = array(rnorm(100*2), c(100,2))
E = array(rnorm(10*2), c(10,2))

df1 = reinflateTensor(A, B, C)
df2 = reinflateTensor(A, D, E)
Y = A[,1]
datasets = list(df1, df2)
modes = list(c(1,2,3), c(1,4,5))
Z = setupCMTFdata(datasets, modes, normalize=FALSE)

# specific setting to reduce runtime for CRAN
result = cv_degeneracy(Z,
                       Y,
                       numComponents=1:3,
                       pis=c(0.50, 0.75),
                       rel_tol=1e-5,
                       abs_tol=1e-5,
                       nstart=1,
                       numCores=1)
}
