% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Georgiou2025}
\alias{Georgiou2025}
\title{Georgiou2025 Apical Periodontitis data}
\format{
\subsection{\code{Georgiou2025}}{

A list object with two elements
\describe{
\item{Inflammatory_mediators}{Longitudinally measured inflammatory mediator data.}
\item{Tooth_microbiome}{Single-timepoint tooth microbiome data of the extracted tooth}
}
}
}
\source{
TBD
}
\usage{
Georgiou2025
}
\description{
The Georgiou longitudinal dataset as a three-dimensional array and a matrix,
with subjects in mode 1, features in mode 2, and time in mode3.
}
\keyword{datasets}
