% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeCMTF.R
\name{initializeCMTF}
\alias{initializeCMTF}
\title{Initialize input vectors for the CMTF algorithm}
\usage{
initializeCMTF(Z, numComponents, initialization = "random", output = "Fac")
}
\arguments{
\item{Z}{List object as generated by \code{\link[=setupCMTFdata]{setupCMTFdata()}}.}

\item{numComponents}{Integer stating the number of desired components for the CMTF model.}

\item{initialization}{Initialization method, either "random" or "nvec" (default "random"). Random will initialize random input vectors. Nvec will initialize vectors according to an singular value decomposition of the (matricized, if needed) concatenated datasets per mode.}

\item{output}{How to return output: as a "Fac" object (default) or vectorized ("vect").}
}
\value{
List or vector of initialized input vectors per mode.
}
\description{
Initialize input vectors for the CMTF algorithm
}
\examples{
I = 108                             # subject mode
J = 100                             # feature mode
K = 10                              # condition mode

df = array(rnorm(I*J*K), c(I,J,K))  # Create the data
datasets = list(df, df)             # Two hypothetical data blocks
modes = list(c(1,2,3), c(1,4,5))    # Shared subject mode, feature and time modes separate
Z = setupCMTFdata(datasets, modes)
initializeCMTF(Z, 1)
}
