% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mize_runner}
\alias{mize_runner}
\title{Runs mize iteratively, allowing for debugging per iteration}
\usage{
mize_runner(
  par,
  fg,
  max_iter = 10000,
  max_fn = 10000,
  abs_tol = 1e-08,
  rel_tol = 1e-08,
  grad_tol = 1e-08,
  cg_update = "HS",
  line_search = "MT"
)
}
\arguments{
\item{par}{Initial parameters}

\item{fg}{Loss and gradient functions in a list (see mize::mize())}

\item{max_iter}{Maximum number of iterations.}

\item{max_fn}{Maximum number of function evaluations.}

\item{abs_tol}{Function tolerance criterion for convergence.}

\item{rel_tol}{Relative function tolerance criterion for convergence.}

\item{grad_tol}{Absolute tolerence for the l2-norm of the gradient vector.}

\item{cg_update}{Update method for the conjugate gradient algorithm, see \code{\link[mize:mize]{mize::mize()}} for the options (default="HS", Hestenes-Steifel).}

\item{line_search}{Line search algorithm to use, see \code{\link[mize:mize]{mize::mize()}} for the options (default="MT", More-Thuente).}
}
\value{
Model object after convergence
}
\description{
Runs mize iteratively, allowing for debugging per iteration
}
\examples{
A = array(rnorm(108*2), c(108, 2))
B = array(rnorm(100*2), c(100, 2))
C = array(rnorm(10*2), c(10, 2))
D = array(rnorm(100*2), c(100,2))
E = array(rnorm(10*2), c(10,2))

df1 = reinflateTensor(A, B, C)
df2 = reinflateTensor(A, D, E)
datasets = list(df1, df2)
modes = list(c(1,2,3), c(1,4,5))
Z = setupCMTFdata(datasets, modes, normalize=FALSE)
numComponents = 1
initialization="random"

init = initializeCMTF(Z, numComponents, initialization, output="vect")
fg = list("fn"=function(x){return(CMTFtoolbox::cmtf_fun(x,Z))},
         "gr"=function(x){return(CMTFtoolbox::cmtf_gradient(x,Z))})
model = mize_runner(init, fg)
}
