% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reinflateMatrix}
\alias{reinflateMatrix}
\title{Create a matrix from a matrix of scores and loadings similar to a component model.}
\usage{
reinflateMatrix(A, B)
}
\arguments{
\item{A}{I x N matrix corresponding to scores for N components.}

\item{B}{J x N matrix corresponding to loadings for N components.}
}
\value{
M, an I x J matrix.
}
\description{
Create a matrix from a matrix of scores and loadings similar to a component model.
}
\examples{
A = rnorm(108)
B = rnorm(100)
M = reinflateMatrix(A,B)
}
