% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalizeFac}
\alias{normalizeFac}
\title{Normalize all vectors in model output Fac object to norm 1.}
\usage{
normalizeFac(Fac)
}
\arguments{
\item{Fac}{List object with all components per mode per item.}
}
\value{
List object of normalized Fac object and the extracted norms per component.
}
\description{
Normalize all vectors in model output Fac object to norm 1.
}
\examples{
A = array(rnorm(108*2), dim(108,2))
B = array(rnorm(100*2), dim(100,2))
C = array(rnorm(10*2), dim(10,2))
Fac = list(A,B,C)
output = normalizeFac(Fac)
}
