% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reinflateFac}
\alias{reinflateFac}
\title{Reinflate all datablocks from a model Fac object.}
\usage{
reinflateFac(Fac, Z, returnAsTensor = FALSE)
}
\arguments{
\item{Fac}{Fac object output from CMTF and ACMTF}

\item{Z}{Z object as generated by \code{\link[=setupCMTFdata]{setupCMTFdata()}}.}

\item{returnAsTensor}{Boolean to return data blocks as rTensor tensor objects (default FALSE)}
}
\value{
List of data blocks
}
\description{
Basically a wrapper function for \code{\link[=reinflateTensor]{reinflateTensor()}} and \code{\link[=reinflateMatrix]{reinflateMatrix()}}.
}
\examples{
I = 108
J = 100
K = 10
df = array(rnorm(I*J*K), c(I,J,K))
datasets = list(df, df)
modes = list(c(1,2,3), c(1,4,5))
Z = setupCMTFdata(datasets, modes)
result = cmtf_opt(Z, 1, max_iter=2)
reinflateFac(result$Fac, Z)
}
