% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setupCMTFdata}
\alias{setupCMTFdata}
\title{Set up datasets for (A)CMTF input}
\usage{
setupCMTFdata(datasets, modes, normalize = TRUE)
}
\arguments{
\item{datasets}{List of arrays of datasets. Multi-way and two-way may be combined.}

\item{modes}{Numbered modes per dataset in a list. Example element 1: 1 2 3 and element 2: 1 4 for the X tensor and Y matrix case with a shared subject mode.}

\item{normalize}{Boolean specifying if the datasets should be normalized to Frobenium norm 1.

Note: this function puts zeroes in positions with missing values.
The indices of missing data are conserved in the output.}
}
\value{
Z, a list with "object" listing the datasets, "sizes" with their size, "norms" with their norms and "missing" stating the missing data.
}
\description{
Set up datasets for (A)CMTF input
}
\examples{
I = 108                             # subject mode
J = 100                             # feature mode
K = 10                              # condition mode

df = array(rnorm(I*J*K), c(I,J,K))  # Create the data
datasets = list(df, df)             # Two hypothetical data blocks
modes = list(c(1,2,3), c(1,4,5))    # Shared subject mode, feature and time modes separate
Z = setupCMTFdata(datasets, modes)
}
