% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vect_to_fac}
\alias{vect_to_fac}
\title{Helper function to convert vectorized output of (a)cmtf to a Fac list object with all loadings per mode.}
\usage{
vect_to_fac(vect, Z)
}
\arguments{
\item{vect}{Vectorized output of (a)cmtf}

\item{Z}{Original Z input object (see \link{setupCMTFdata}).}
}
\value{
Fac: list object with all loadings in all components per mode, ordered the same way as Z$modes.
}
\description{
Helper function to convert vectorized output of (a)cmtf to a Fac list object with all loadings per mode.
}
\examples{
A = array(rnorm(108*2), c(108, 2))
B = array(rnorm(100*4), c(100, 4))
C = array(rnorm(10*4), c(10, 4))

df1 = reinflateTensor(A, B[,1:2], C[,1:2])
df2 = reinflateTensor(A, B[,3:4], C[,3:4])
datasets = list(df1, df2)
modes = list(c(1,2,3), c(1,4,5))
Z = setupCMTFdata(datasets, modes)
result = cmtf_opt(Z, 2, initialization="random", max_iter = 2)
Fac = vect_to_fac(result$par)
}
